backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.10.3
pandas==2.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
email-validator==2.0.0
numpy==1.24.3
